requirements.txt
asttokens==2.0.8
backcall==0.2.0
breadability==0.1.20
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==0.10.0
filelock==3.8.0
huggingface-hub==0.9.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
lxml==4.9.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.2
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pycountry==22.3.5
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
sentencepiece==0.1.97
six==1.16.0
stack-data==0.4.0
sumy==0.10.0
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
